requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
Babel==2.9.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.2
currencies==2020.12.12
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.3
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-celery-results==2.3.1
django-coverage==1.2.4
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-money==2.1.1
django-redis==5.2.0
django-tables2==2.4.1
django-tablib==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
Faker==13.3.4
idna==3.3
kombu==5.2.4
MarkupPy==1.14
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pika==1.2.0
pika-stubs==0.1.3
Pillow==9.0.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
py-moneyed==1.2
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
tablib==3.2.1
tabulate==0.8.9
tzdata==2021.5
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
